Extract Library + Vinted scanners into services
Move the two inline HTML scanners out of SyncManager so server.ts stays a
thin orchestration root. The Library (OPAC MBP Lublin) and Vinted scraping
loops now live in LibraryCheckService and VintedSyncService, each exposing
run*(sendEvent, checkCancellation) like every other sync service; SyncManager
dispatches them via executeTask. Behavior is unchanged.

- New scrapingClient.ts (repo root) holds the shared User-Agent rotation and
  keep-alive HTTPS agent both scanners need.
- Route the scanners' console.* through the structured logger.
- Drop now-dead imports from server.ts (axios, https, withRetry, cleanTitle,
  calculateSimilarity, normalizeAuthor, WikiParser, p-limit); server.ts shrinks
  from 741 to ~430 lines.
- Add focused tests for both services (candidate filtering, match emission,
  no-results, cancellation) with axios mocked.
- Update guidelines, CLAUDE.md map, and the two scanner docs to point at the
  new service locations.
