+++ COGITATOR_GUIDELINES.md
@@ -3,7 +3,7 @@
 ## 1. CORE ARCHITECTURE (BACKEND)
 - **Pattern**: Service-Adapter-Manager.
 - **Adapters**: `NotionAdapter`, `WikiAdapter`. Pure API wrappers. No business logic.
-- **Services**: `*SyncService`, `StatsService`, `IntegrityService`, `PurificationService`, `SchemaValidationService`. Logic-heavy, stateless where possible.
+- **Services**: `*SyncService`, `StatsService`, `IntegrityService`, `PurificationService`, `SchemaValidationService`, plus the HTML scanners `LibraryCheckService` and `VintedSyncService`. Logic-heavy, stateless where possible. `SyncManager` stays a thin dispatcher — every long-running ritual lives in a service exposing `run*(sendEvent, checkCancellation)`; never inline scraping/sync loops back into `server.ts`. Scanners scrape public HTML (not Notion/Wiki APIs) and share `scrapingClient.ts` (User-Agent rotation + keep-alive HTTPS agent).
 - **Orchestration**: `SyncManager` (in `server.ts`). Each task owns a `SyncTask` state object with its own cancellation flag; `executeTask` releases the lock only if it still belongs to the finishing task, and `resetSyncState` cancels the orphaned task before releasing the lock. Never reintroduce shared mutable booleans for task state.
 - **Communication**: SSE (Server-Sent Events) for real-time progress. Use `sendEvent({ type, ... })`. Client-disconnect cancellation MUST listen on `res.on("close")`, NOT `req.on("close")` — for a POST with a body, `req` emits `close` when `express.json()` finishes reading the body (mid-response), which spuriously cancelled active syncs. Guard writes with `writableEnded`.
 - **SSE hosting hardening**: proxies (e.g. Render) buffer streamed responses. `setupSSE` sends `X-Accel-Buffering: no`, `flushHeaders()`, a ~2KB comment padding to push past the buffer threshold, and a 5s keepalive. The client (`useSync`) has a 30s stall watchdog. Do not remove these.
