+++ docs/vinted-scanner.md
@@ -3,7 +3,7 @@
 ## 1. Overview
 Searches **vinted.pl** for physical, second-hand copies of the tracked books. This is a direct HTML scraper — it fetches Vinted's catalog search page over HTTP and parses the response. **It does not use any AI / LLM.** (An earlier design intended a Gemini-based search; that was never implemented, and the `@google/genai` dependency has been removed. `GEMINI_API_KEY` remains only as an unused/reserved env var.)
 
-## 2. Search Logic (`checkVintedAvailability` in `server.ts`)
+## 2. Search Logic (`VintedSyncService.runVintedCheck` in `services/vintedSyncService.ts`)
 - **Candidates**: Books from Notion that are NOT already read/owned (excludes the `Przeczytane`, `Biblioteka`, `Biblioteka 9`, `Posiadam` source tags), with a non-empty Polish title.
 - **Search URL**: `https://www.vinted.pl/catalog?catalog[]=2319&language_book_ids[]=6440&page=1&order=price_low_to_high&price_from=2&currency=PLN&search_text={title author}` (books category, cheapest first).
 - **Request shaping**: Rotates a realistic browser `User-Agent` and sends browser-like headers (`Referer`, `Sec-Fetch-*`, etc.) to reduce bot blocking.
